# Conflicts:
#	src/main/java/com/example/idatt2106_2022_05_backend/config/SecurityConfiguration.java
#	src/main/java/com/example/idatt2106_2022_05_backend/config/WebSocketConfig.java
#	src/main/resources/application.yml

+++ src/main/java/com/example/idatt2106_2022_05_backend/controller/UserController.java
@@ -14,6 +14,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 
@@ -39,9 +40,32 @@ public class UserController {
         return userService.updateUser(userId, userUpdateDto);
     }
 
+    @PutMapping("/profilePicture/{userId}")
+    @ApiOperation(value = "Endpoint to update user profile picture", response = Response.class)
+    public Response updatePicture(@PathVariable Long userId, @RequestPart MultipartFile file) throws IOException {
+        log.debug("[X] Call to update user with id = {}", userId);
+        if(!securityService.isUser(userId)){
+            return new Response("Du har ikke tilgang", HttpStatus.BAD_REQUEST);
+        }
+        return userService.updatePicture(userId, file);
+    }
+
+    @GetMapping("/profilePicture/{userId}")
+    @ApiOperation(value = "Endpoint to update user profile picture", response = Response.class)
+    public Response updatePicture(@PathVariable Long userId) throws IOException {
+        log.debug("[X] Call to update user with id = {}", userId);
+        if(!securityService.isUser(userId)){
+            return new Response("Du har ikke tilgang", HttpStatus.BAD_REQUEST);
+        }
+        return userService.getPicture(userId);
+    }
+
     @DeleteMapping("/profilePicture")
     @ApiOperation(value = "Endpoint to delete profile picture", response = Response.class)
     public Response deleteProfilePicture(@ModelAttribute UpdatePictureDto updatePictureDto) {
+        if(!securityService.userPicture(updatePictureDto.getId(), updatePictureDto.getUserId())){
+            return new Response("Du har ikke tilgang", HttpStatus.BAD_REQUEST);
+        }
         try {
             return userService.deleteProfilePicture(updatePictureDto.getUserId(),
                     updatePictureDto.getMultipartFile().getBytes());
@@ -64,8 +88,8 @@ public class UserController {
     @GetMapping("/{userId}")
     @ApiOperation(value = "Endpoint to get user", response = UserReturnDto.class)
     public Response getUser(@PathVariable Long userId, Authentication auth) {
-        UserDetails user = (UserDetails)auth.getPrincipal();
-        System.out.println(user.getUsername());
+//        UserDetails user = (UserDetails)auth.getPrincipal();
+//        System.out.println(user.getUsername());
         log.debug("[X] Call to get user with id = {}", userId);
         return userService.getUser(userId);
     }
